Refactor: rename rex to expr
